refactor: Switch from layered architecture to feature (package feature.parkify)
