+++ src/main/java/com/igrowker/feature/parkify/MiniProjectApplication.java
@@ -1,0 +1,14 @@
+package com.igrowker.feature.parkify;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
+
+@SpringBootApplication(exclude = {SecurityAutoConfiguration.class})
+public class MiniProjectApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(MiniProjectApplication.class, args);
+	}
+
+}
