+++ src/main/java/com/igrowker/feature/parkify/controllers/BookingController.java
@@ -1,0 +1,12 @@
+package com.igrowker.feature.parkify.controllers;
+
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@CrossOrigin(origins = "")
+@RestController
+@RequestMapping("/api/bookings")
+public class BookingController {
+
+}
