+++ src/main/java/com/igrowker/feature/parkify/controllers/HelloWorldController.java
@@ -1,0 +1,13 @@
+package com.igrowker.feature.parkify.controllers;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class HelloWorldController {
+
+    @GetMapping("/hello")
+    public String sayHello() {
+        return "Hello, World!";
+    }
+}
