+++ src/main/java/com/igrowker/feature/parkify/controllers/ParkingSpotController.java
@@ -1,0 +1,4 @@
+package com.igrowker.feature.parkify.controllers;
+
+public class ParkingSpotController {
+}
