+++ src/main/java/com/igrowker/feature/parkify/controllers/UserController.java
@@ -1,0 +1,11 @@
+package com.igrowker.feature.parkify.controllers;
+
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@CrossOrigin(origins = "")
+@RestController
+@RequestMapping("/api/users")
+public class UserController {
+}
