+++ src/main/java/com/igrowker/feature/parkify/excepcions/ResourceNotFoundException.java
@@ -1,0 +1,4 @@
+package com.igrowker.feature.parkify.excepcions;
+
+public class ResourceNotFoundException {
+}
