+++ src/main/java/com/igrowker/feature/parkify/models/Parking.java
@@ -1,0 +1,17 @@
+package com.igrowker.feature.parkify.models;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@Entity
+@Table(name="parking")
+public class Parking {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+}
