+++ src/main/java/com/igrowker/feature/parkify/models/ParkingHistory.java
@@ -1,0 +1,4 @@
+package com.igrowker.feature.parkify.models;
+
+public class ParkingHistory {
+}
