+++ src/main/java/com/igrowker/feature/parkify/models/ParkingSpot.java
@@ -1,0 +1,15 @@
+package com.igrowker.feature.parkify.models;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.Table;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@Entity
+@Table(name="parkingSpot")
+public class ParkingSpot {
+}
