+++ src/main/java/com/igrowker/feature/parkify/repositories/IParkingRepository.java
@@ -1,0 +1,7 @@
+package com.igrowker.feature.parkify.repositories;
+
+import com.igrowker.feature.parkify.models.Parking;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface IParkingRepository extends JpaRepository<Parking, Long> {
+}
