+++ src/main/java/com/igrowker/feature/parkify/repositories/IUserRepository.java
@@ -1,0 +1,7 @@
+package com.igrowker.feature.parkify.repositories;
+
+import com.igrowker.feature.parkify.models.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface IUserRepository extends JpaRepository<User, Long> {
+}
