+++ src/main/java/com/igrowker/feature/parkify/services/ParkingService.java
@@ -1,0 +1,7 @@
+package com.igrowker.feature.parkify.services;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class ParkingService {
+}
