+++ src/main/java/com/igrowker/feature/parkify/services/ParkingSpotService.java
@@ -1,0 +1,4 @@
+package com.igrowker.feature.parkify.services;
+
+public class ParkingSpotService {
+}
