--- src/main/java/com/igrowker/miniproject/MiniProjectApplication.java
@@ -1,14 +1,0 @@
-package com.igrowker.miniproject;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
-
-@SpringBootApplication(exclude = {SecurityAutoConfiguration.class})
-public class MiniProjectApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(MiniProjectApplication.class, args);
-	}
-
-}
