--- src/main/java/com/igrowker/miniproject/controllers/BookingController.java
@@ -1,12 +1,0 @@
-package com.igrowker.miniproject.controllers;
-
-import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@CrossOrigin(origins = "")
-@RestController
-@RequestMapping("/api/bookings")
-public class BookingController {
-
-}
