--- src/main/java/com/igrowker/miniproject/controllers/HelloWorldController.java
@@ -1,13 +1,0 @@
-package com.igrowker.miniproject.controllers;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class HelloWorldController {
-
-    @GetMapping("/hello")
-    public String sayHello() {
-        return "Hello, World!";
-    }
-}
