--- src/main/java/com/igrowker/miniproject/controllers/ParkingSpotController.java
@@ -1,4 +1,0 @@
-package com.igrowker.miniproject.controllers;
-
-public class ParkingSpotController {
-}
