--- src/main/java/com/igrowker/miniproject/controllers/UserController.java
@@ -1,11 +1,0 @@
-package com.igrowker.miniproject.controllers;
-
-import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@CrossOrigin(origins = "")
-@RestController
-@RequestMapping("/api/users")
-public class UserController {
-}
