--- src/main/java/com/igrowker/miniproject/excepcions/ResourceNotFoundException.java
@@ -1,4 +1,0 @@
-package com.igrowker.miniproject.excepcions;
-
-public class ResourceNotFoundException {
-}
