--- src/main/java/com/igrowker/miniproject/models/ParkingHistory.java
@@ -1,4 +1,0 @@
-package com.igrowker.miniproject.models;
-
-public class ParkingHistory {
-}
