--- src/main/java/com/igrowker/miniproject/models/ParkingSpot.java
@@ -1,15 +1,0 @@
-package com.igrowker.miniproject.models;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.Table;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@Entity
-@Table(name="parkingSpot")
-public class ParkingSpot {
-}
