--- src/main/java/com/igrowker/miniproject/models/User.java
@@ -1,17 +1,0 @@
-package com.igrowker.miniproject.models;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@Entity
-@Table(name= "user")
-public class User {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-}
