--- src/main/java/com/igrowker/miniproject/services/ParkingSpotService.java
@@ -1,4 +1,0 @@
-package com.igrowker.miniproject.services;
-
-public class ParkingSpotService {
-}
