--- src/main/java/com/igrowker/miniproject/services/UserService.java
@@ -1,7 +1,0 @@
-package com.igrowker.miniproject.services;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class UserService {
-}
