+++ src/test/java/com/igrowker/feature/parkify/MiniProjectApplicationTests.java
@@ -1,0 +1,7 @@
+package com.igrowker.feature.parkify;
+
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class MiniProjectApplicationTests {
+}
