+++ src/test/java/com/igrowker/feature/parkify/controllers/UserControllerTest.java
@@ -1,0 +1,5 @@
+package com.igrowker.feature.parkify.controllers;
+
+class UserControllerTest {
+
+}
