+++ src/test/java/com/igrowker/feature/parkify/services/UserServiceTest.java
@@ -1,0 +1,5 @@
+package com.igrowker.feature.parkify.services;
+
+class UserServiceTest {
+
+}
