--- src/test/java/com/igrowker/miniproject/MiniProjectApplicationTests.java
@@ -1,8 +1,0 @@
-package com.igrowker.miniproject;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class MiniProjectApplicationTests {
-}
