--- src/test/java/com/igrowker/miniproject/controllers/UserControllerTest.java
@@ -1,7 +1,0 @@
-package com.igrowker.miniproject.controllers;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class UserControllerTest {
-
-}
