--- src/test/java/com/igrowker/miniproject/services/UserServiceTest.java
@@ -1,7 +1,0 @@
-package com.igrowker.miniproject.services;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class UserServiceTest {
-
-}
